completed cart and small fixes
