+++ db.json
@@ -268,6 +268,8 @@
       "img": "sauces/brown.webp"
     }
   ],
-  "selectedItems": [],
+  "selectedItems": [
+
+  ],
   "users": []
 }
